pyproject.toml
[tool.poetry]
name = "aoc23"
version = "0.1.0"
description = ""
authors = ["Tony ElHabr <anthonyelhabr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
advent-of-code-data = {extras = ["nb"], version = "^2.0.1"}
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
